api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
mysql-connector-python==8.3.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
pydantic==2.6.1
